Remove entity list
